:tada: Duplicate token group - tests
